* Adding s390x arch support to the k8s-1.28 provider in SLIM mode * Used qemu-monitor command to add secondary network interface to avoid the primary network interface having names other than eth0. As  based on the MAC address of eth0 interface we are fetching the IP of the VM from DHCP server, it needs to be MAC of the primary network interface, to obtain the IP for the interface. * Bumped calico CNI version from 3.18.0 to 3.27.2 (latest currently), as 3.18.0 doesn't support s390x arch.

+++ cluster-provision/centos9/scripts/download_box.sh
@@ -3,6 +3,14 @@
 set -e
 set -o pipefail
 
-curl -L $1 | tar -zxvf - box.img
-qemu-img convert -O qcow2 box.img box.qcow2
-rm box.img
+
+ARCH=$(uname -m)
+
+#For the s390x architecture, instead of vagrant box image, generic cloud (qcow2) image is used directly.
+if [ "$ARCH" == "s390x" ]; then
+    curl -L $1 -o box.qcow2
+else
+    curl -L $1 | tar -zxvf - box.img
+    qemu-img convert -O qcow2 box.img box.qcow2
+    rm box.img
+fi
